requirements.txt
# Core dependencies
python-telegram-bot>=20.0
python-dotenv>=1.0.0

# Logging
colorlog>=6.7.0

# HTTP requests
requests>=2.31.0

# AI/ML (optional, uncomment as needed)
# openai>=1.0.0
# anthropic>=0.3.0
